+ Validaciones
Se agregan validaciones para el campo de contraseña del formulario de Registro de clientes. Se acomoda mejor la estructura de carpetas de la vistas
