+++ TemplateTPIntegrador/TemplateTPIntegrador/MenuForm.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace TemplateTPIntegrador
+{
+    public partial class MenuForm : Form
+    {
+        public MenuForm()
+        {
+            InitializeComponent();
+        }
+
+        private void linkLabel4_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
+        {
+
+        }
+    }
+}
